Refactor code to use Bazel Build
