+++ assets/BUILD
@@ -1,0 +1,8 @@
+load("@rules_cc//cc:defs.bzl", "cc_binary", "cc_library")
+
+cc_library (
+    name = "assets",
+    srcs = glob(["*.cc"]),
+    hdrs = glob(["*.h"]),
+    visibility = ["//visibility:public"],
+)
