+++ assets/asset_enums.h
@@ -1,0 +1,3 @@
+enum class TickerType{
+    isin, cusip
+};
