--- interface.cpp
@@ -1,3 +1,0 @@
-using namespace std;
-
-#include "asset.h"
