+++ pricers/derivatives.hpp
@@ -1,4 +1,4 @@
-#include "unit structures/asset.h"
+#include "assets/asset.h"
 
 class Derivative {
 public:
